+++ FretFramework_CS/Song/Tracks/Notes/Drums/Drums_Legacy.cs
@@ -81,7 +81,7 @@ namespace Framework.Song.Tracks.Notes.Drums
 
         public Drum_Legacy() { }
 
-        public bool Set_From_Chart(nuint lane, ulong length)
+        public bool Set_From_Chart(uint lane, ulong length)
         {
             fixed (DrumPad_Pro* pads = &snare)
             {
